requirements.txt
streamlit==1.34.0
pyyaml==6.0.1
onnxruntime==1.18.0
matplotlib>=3.2.2
numpy>=1.18.5
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
Pillow>=7.1.2
pandas>=1.1.4
seaborn>=0.11.0
opencv-python>=4.1.1
gdown==5.2.0